feat: Add simpler schema and build size script
This commit introduces a simpler schema for validation and a new script
to build and show the size of the bundle. The simpler schema is added to
improve the performance of the validation process. The build size script
is added to the package.json file to allow developers to easily check
the size of the bundle after building.